// agent_ihub_web/src/api/types.gen.ts
export enum type {
  MIT = 'MIT',
  APACHE_2_0 = 'APACHE_2_0',
  GPL_3_0 = 'GPL_3_0',
  BSD_3_CLAUSE = 'BSD_3_CLAUSE',
  AGPL_3_0 = 'AGPL_3_0',
  PROPRIETARY = 'PROPRIETARY',
  CUSTOM = 'CUSTOM',
}
// ...
export type FileDocMetadataDTO = {
  name?: string;
  separator?: string;
};

/**
 * 文件信息DTO
 */
export type FileInfoDTO = {
  /**
   * 文件ID
   */
  id?: string;
  /**
   * 完整文件名
   */
  fileName?: string;
  /**
   * 文件大小 (字节)
   */
  fileSize?: number;
  /**
   * 文件类型
   */
  fileType?: string;
  /**
   * 上传类型
   */
  type?: 'TOOLS' | 'AVATARS' | 'KNOWLEDGE';
  /**
   * md5
   */
  md5?: string;
  /**
   * 文档文件的元数据 从liteagent导入的需保留原数据，否则导出时生成
   */
  metadata?: FileDocMetadataDTO;
  /**
   * 关联图片文件ID
   */
  imagesFileIds?: Array<string>;
  /**
   * 上传用户ID
   */
  userId?: string;
  /**
   * 上传时间
   */
  uploadTime?: Date;
};
// ...
export type ConversionProgressDTO = {
  /**
   * 进度
   */
  progress?: number;
  /**
   * 状态
   */
  stage?: string;
  /**
   * 详情
   */
  detail?: string;
  /**
   * 文件信息,完成时返回
   */
  fileInfoDTO?: FileInfoDTO;
};

export type RConversionProgressDTO = {
  code?: number;
  msg?: string;
  data?: ConversionProgressDTO;
};